+++ client/src/App.js
@@ -9,7 +9,6 @@ import EditUserPage from "./pages/EditUserPage";
 import ProfilePage from "./pages/ProfilePage";
 import { UserContextProvider } from "./contexts/userContext";
 import FileSubmitPage from "./pages/FileSubmitPage";
-import FileUploadPage from "./pages/FileUploadPage";
 
 function App() {
   return (
@@ -23,12 +22,10 @@ function App() {
           <Route path="" element={<TeamsPage />} />
           <Route path=":id" element={<DetailedTeamPage />} />
         </Route>
-        <Route path="/editUser" element={<ProfilePage />}>
-          <Route path=":id" element={<EditUserPage />} />
-        </Route>
         <Route path="/my/profile/">
           <Route path=":id" element={<ProfilePage />} />
           <Route path="edit/:id" element={<EditUserPage />} />
+          <Route path="submit/:id" element={<FileSubmitPage />} />
         </Route>
         <Route path="/submit" element={<FileSubmitPage />} />
       </Routes>
