+++ client/src/components/ProfileAppBar.js
@@ -4,9 +4,15 @@ import Box from "@mui/material/Box";
 import Toolbar from "@mui/material/Toolbar";
 import Typography from "@mui/material/Typography";
 import { Link } from "react-router-dom";
-import { Avatar } from "@mui/material";
+import { Avatar, Button } from "@mui/material";
+import { useParams } from "react-router-dom";
+import { useContext } from "react";
+import { UserContext } from "../contexts/userContext";
 
-function ProfileAppBar() {
+function ProfileAppBar(navItems) {
+  const { user, setUser } = useContext(UserContext);
+  console.log(user._id);
+  const nav = navItems["navItems"];
   return (
     <Box sx={{ display: "flex" }}>
       <AppBar
@@ -22,9 +28,23 @@ function ProfileAppBar() {
             component="div"
             sx={{ flexGrow: 1, display: { xs: "none", sm: "block" }, ml: 2 }}
           >
-            <Link to="">FFP Tracker for TFF</Link>
+            <Link to={`/my/profile/${user._id}`}>FFP Tracker for TFF</Link>
           </Typography>
-          <Link to="/login">LOGOUT</Link>
+          <Box sx={{ display: { xs: "none", sm: "block" } }}>
+            {nav.map((item) => (
+              <Button key={item} sx={{ color: "#FFFFFF" }}>
+                <Link to={`/${item.toLowerCase().replace(" ", "")}`}>
+                  {item}
+                </Link>
+              </Button>
+            ))}
+          </Box>
+          <Button sx={{ color: "#FFFFFF" }}>
+            <Link to={`/my/profile/edit/${user._id}`}>Edit Your Profile</Link>
+          </Button>
+          <Button sx={{ color: "#FFFFFF" }}>
+            <Link to="/login">Log Out</Link>
+          </Button>
         </Toolbar>
       </AppBar>
     </Box>
