+++ client/src/pages/ProfilePage.js
@@ -4,7 +4,7 @@ import ProfilePageComponent from "../components/ProfilePageComponent";
 export default function ProfilePage() {
   return (
     <>
-      <ProfileAppBar />
+      <ProfileAppBar navItems={[]} />
       <ProfilePageComponent />
     </>
   );
